fix: palette fixes for more than max short block states

+++ modules/terraform/src/main/java/net/hollowcube/terraform/buffer/palette/FixedPalette.java
@@ -7,7 +7,7 @@ public record FixedPalette(int value) implements Palette {
     @Override
     public @Nullable Block get(int x, int y, int z) {
         if (value == Palette.UNSET) return null;
-        return Block.fromStateId((short) value);
+        return Block.fromStateId(value);
     }
 
     @Override
